refactor: BottomSheet components into separate files to unit test
