+++ app/components/Views/AddAsset/AddAsset.tsx
@@ -5,7 +5,7 @@ import React, {
   useRef,
   useState,
 } from 'react';
-import { SafeAreaView, ScrollView, View } from 'react-native';
+import { SafeAreaView, View } from 'react-native';
 import { useSelector } from 'react-redux';
 import DefaultTabBar from 'react-native-scrollable-tab-view/DefaultTabBar';
 import AddCustomToken from '../../UI/AddCustomToken';
@@ -47,24 +47,12 @@ import { AddAssetParams } from './AddAsset.types';
 import Routes from '../../../constants/navigation/Routes';
 import { NFT_TITLE, TOKEN, TOKEN_TITLE } from './AddAsset.constants';
 import { AddAssetViewSelectorsIDs } from '../../../../e2e/selectors/wallet/AddAssetView.selectors';
-import BottomSheet, {
-  BottomSheetRef,
-} from '../../../component-library/components/BottomSheets/BottomSheet';
-import { VerticalAlignment } from '../../../component-library/components/List/ListItem';
-import ListItemSelect from '../../../component-library/components/List/ListItemSelect';
-import { WalletViewSelectorsIDs } from '../../../../e2e/selectors/wallet/WalletView.selectors';
-import Cell, {
-  CellVariant,
-} from '../../../component-library/components/Cells/Cell';
-import {
-  AvatarSize,
-  AvatarVariant,
-} from '../../../component-library/components/Avatars/Avatar';
-import { getNetworkImageSource } from '../../../util/networks';
+import { BottomSheetRef } from '../../../component-library/components/BottomSheets/BottomSheet';
 import { Hex } from '@metamask/utils';
 import { enableAllNetworksFilter } from '../../UI/Tokens/util/enableAllNetworksFilter';
 import Engine from '../../../core/Engine';
-import NetworkImageComponent from '../../UI/NetworkImages';
+import NetworkListBottomSheet from './components/NetworkListBottomSheet';
+import NetworkFilterBottomSheet from './components/NetworkFilterBottomSheet';
 
 export enum FilterOption {
   AllNetworks,
@@ -186,111 +174,24 @@ const AddAsset = () => {
 
   const renderNetworkSelector = useCallback(
     () => (
-      <BottomSheet
-        shouldNavigateBack={false}
-        ref={sheetRef}
-        onClose={() => setOpenNetworkSelector(false)}
-        isInteractable
-        style={styles.bottomSheetWrapperContent}
-      >
-        <Text variant={TextVariant.HeadingMD} style={styles.bottomSheetTitle}>
-          {strings('networks.select_network')}
-        </Text>
-        <ScrollView>
-          {Object.values(networkConfigurations).map((network) => (
-            <View style={styles.bottomSheetWrapper} key={network.chainId}>
-              <Cell
-                variant={CellVariant.Select}
-                title={network.name}
-                avatarProps={{
-                  variant: AvatarVariant.Network,
-                  name: network.name,
-                  // @ts-expect-error - The utils/network file is still JS and this function expects a networkType, and should be optional
-                  imageSource: getNetworkImageSource({
-                    chainId: network.chainId,
-                  }),
-                  size: AvatarSize.Sm,
-                }}
-                onPress={() => {
-                  setSelectedNetwork(network.chainId);
-                  setOpenNetworkSelector(false);
-                }}
-                isSelected={selectedNetwork === network.chainId}
-              />
-            </View>
-          ))}
-        </ScrollView>
-      </BottomSheet>
+      <NetworkListBottomSheet
+        selectedNetwork={selectedNetwork}
+        setSelectedNetwork={setSelectedNetwork}
+        setOpenNetworkSelector={setOpenNetworkSelector}
+        sheetRef={sheetRef}
+      />
     ),
-
     // eslint-disable-next-line react-hooks/exhaustive-deps
     [openNetworkSelector, networkConfigurations, selectedNetwork],
   );
 
   const renderNetworkFilterSelector = useCallback(
     () => (
-      <BottomSheet
-        shouldNavigateBack={false}
-        ref={sheetRef}
-        onClose={() => setOpenNetworkFilter(false)}
-        isInteractable
-      >
-        <View style={styles.bottomSheetWrapper}>
-          <Text variant={TextVariant.HeadingMD} style={styles.bottomSheetTitle}>
-            {strings('wallet.filter_by')}
-          </Text>
-          <ListItemSelect
-            testID={WalletViewSelectorsIDs.TOKEN_NETWORK_FILTER_ALL}
-            onPress={() => {
-              onFilterControlsBottomSheetPress(FilterOption.AllNetworks);
-              setOpenNetworkFilter(false);
-            }}
-            isSelected={
-              isAllNetworksEnabled &&
-              Object.keys(isAllNetworksEnabled).length > 1
-            }
-            gap={8}
-            verticalAlignment={VerticalAlignment.Center}
-          >
-            <Text style={styles.bottomSheetText}>
-              {`${strings('app_settings.popular')} ${strings(
-                'app_settings.networks',
-              )}`}
-            </Text>
-            <View style={styles.networkImageContainer}>
-              <NetworkImageComponent
-                isAllNetworksEnabled
-                allNetworksEnabled={allNetworksEnabled}
-                selectorButtonDisplayed={false}
-              />
-            </View>
-          </ListItemSelect>
-          <ListItemSelect
-            testID={WalletViewSelectorsIDs.TOKEN_NETWORK_FILTER_CURRENT}
-            onPress={() => {
-              onFilterControlsBottomSheetPress(FilterOption.CurrentNetwork);
-              setOpenNetworkFilter(false);
-            }}
-            isSelected={
-              isAllNetworksEnabled &&
-              Object.keys(isAllNetworksEnabled).length === 1
-            }
-            gap={8}
-            verticalAlignment={VerticalAlignment.Center}
-          >
-            <Text style={styles.bottomSheetText}>
-              {strings('wallet.current_network')}
-            </Text>
-            <View style={styles.networkImageContainer}>
-              <NetworkImageComponent
-                isAllNetworksEnabled={false}
-                allNetworksEnabled={{ [chainId]: true }}
-                selectorButtonDisplayed={false}
-              />
-            </View>
-          </ListItemSelect>
-        </View>
-      </BottomSheet>
+      <NetworkFilterBottomSheet
+        onFilterControlsBottomSheetPress={onFilterControlsBottomSheetPress}
+        setOpenNetworkFilter={setOpenNetworkFilter}
+        sheetRef={sheetRef}
+      />
     ),
     // eslint-disable-next-line react-hooks/exhaustive-deps
     [allNetworksEnabled, onFilterControlsBottomSheetPress],
